+++ pyannote/audio/core/model.py
@@ -35,8 +35,8 @@ import torch.nn as nn
 import torch.optim
 from huggingface_hub import hf_hub_download
 from huggingface_hub.utils import RepositoryNotFoundError
+from lightning_fabric.utilities.cloud_io import _load as pl_load
 from pyannote.core import SlidingWindow
-from lightning_lite.utilities.cloud_io import _load as pl_load
 from pytorch_lightning.utilities.model_summary import ModelSummary
 from semver import VersionInfo
 from torch.utils.data import DataLoader
@@ -522,9 +522,6 @@ class Model(pl.LightningModule):
     # model does not really need to know how it is being used.
     def validation_step(self, batch, batch_idx):
         return self.task.validation_step(batch, batch_idx)
-
-    def validation_epoch_end(self, outputs):
-        return self.task.validation_epoch_end(outputs)
 
     def configure_optimizers(self):
         return torch.optim.Adam(self.parameters(), lr=1e-3)