+++ pyannote/audio/pipelines/__init__.py
@@ -24,13 +24,11 @@ from .multilabel import MultiLabelSegmentation
 from .overlapped_speech_detection import OverlappedSpeechDetection
 from .resegmentation import Resegmentation
 from .speaker_diarization import SpeakerDiarization
-from .speaker_segmentation import SpeakerSegmentation
 from .voice_activity_detection import VoiceActivityDetection
 
 __all__ = [
     "VoiceActivityDetection",
     "OverlappedSpeechDetection",
-    "SpeakerSegmentation",
     "SpeakerDiarization",
     "Resegmentation",
     "MultiLabelSegmentation",