+++ pyannote/audio/utils/powerset.py
@@ -24,11 +24,7 @@
 # Hervé BREDIN - https://herve.niderb.fr
 # Alexis PLAQUET
 
-try:
-    from functools import cached_property
-except ImportError:
-    from backports.cached_property import cached_property
-
+from functools import cached_property
 from itertools import combinations
 
 import scipy.special